+++ src/components/Card/__tests__/Card.spec.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { mount } from 'enzyme';
 
 import Card from '../Card';
+import { Icon } from '../../..';
 
 describe('<Card />', () => {
   let mounted = null;
@@ -14,11 +15,11 @@ describe('<Card />', () => {
 
   beforeEach(() => {
     props = {
-      icon: 'contact',
-      sprite: 'standard',
+      boundary: true,
+      icon: <Icon icon="contact" sprite="standard" />,
       title: 'Base Card',
       headerRight: 'Right Header',
-      body: 'Body',
+      children: 'Body',
       footer: 'footer',
     };
 
@@ -27,7 +28,7 @@ describe('<Card />', () => {
   });
 
   it('renders the icon', () => {
-    expect(mounted.find('svg').length).toEqual(1);
+    expect(mounted.find(Icon).exists()).toBeTruthy();
   });
 
   it('renders the header', () => {
@@ -39,12 +40,17 @@ describe('<Card />', () => {
   });
 
   it('renders body', () => {
-    expect(mounted.find('div.slds-card__body').first().text()).toEqual(props.body);
+    expect(mounted.find('div.slds-card__body').first().text()).toEqual(props.children);
   });
 
   it('renders footer', () => {
     expect(mounted.find('footer.slds-card__footer').first().text()).toEqual(props.footer);
   });
+
+  it('allows passing boundary flag', () => {
+    expect(mounted.find('.slds-card').hasClass('slds-card_boundary')).toBeTruthy();
+  });
+
 
   it('applies className and rest-properties', () => {
     mounted.setProps({ className: 'foo', 'data-test': 'bar' });